+++ c/dsls/Makefile.no-autoconfig
@@ -112,6 +112,8 @@ lemon: lemon.c \
   lemon_error.c \
   lemon_string.h \
   lemon_string.c \
+  lemon_msort.h \
+  lemon_msort.c \
   lemon_set.h \
   lemon_set.c \
   lemon_memory.h \
@@ -123,13 +125,15 @@ lemon: lemon.c \
   lemon_action.c \
   lemon_symbol.h \
   lemon_symbol.c \
+  lemon_config_list.h \
+  lemon_config_list.c \
   lemon_plink.h \
   lemon_plink.c \
   lemon_parse.h \
   lemon_parse.c \
   lemon_report.h \
   lemon_report.c
-	$(DSLCC) -o lemon lemon.c lemon_assert.c lemon_error.c lemon_string.c lemon_set.c lemon_memory.c lemon_option.c lemon_action.c lemon_symbol.c lemon_plink.c lemon_parse.c lemon_report.c
+	$(DSLCC) -o lemon lemon.c lemon_assert.c lemon_error.c lemon_string.c lemon_msort.c lemon_set.c lemon_memory.c lemon_option.c lemon_action.c lemon_symbol.c lemon_config_list.c lemon_plink.c lemon_parse.c lemon_report.c
 
 # ----------------------------------------------------------------
 clean:
