+++ c/dsls/lemon_config_list.h
@@ -1,0 +1,24 @@
+#ifndef LEMON_CONFIG_LIST_H
+#define LEMON_CONFIG_LIST_H
+
+#include "lemon_structs.h"
+
+int Configcmp(struct config *, struct config *);
+
+void Configlist_init(void);
+struct config *Configlist_add(struct rule *, int);
+struct config *Configlist_addbasis(struct rule *, int);
+void Configlist_closure(struct lemon*);
+void Configlist_sort(void);
+void Configlist_sortbasis(void);
+struct config *Configlist_return(void);
+struct config *Configlist_basis(void);
+void Configlist_eat(struct config *);
+void Configlist_reset(void);
+
+void Configtable_init(void);
+int Configtable_insert(struct config *);
+struct config *Configtable_find(struct config *);
+void Configtable_clear(int(*)(struct config *));
+
+#endif // LEMON_CONFIG_LIST_H
