+++ c/dsls/lemon_msort.h
@@ -1,0 +1,7 @@
+#ifndef LEMON_MSORT_H
+#define LEMON_MSORT_H
+
+// xxx full proto
+char *msort();
+
+#endif // LEMON_MSORT_H
